Fixed bugs

+++ lib/docman/commands/clean_changed_cmd.rb
@@ -11,8 +11,14 @@ module Docman
     def execute
       if File.directory? @context['full_build_path']
         if @context.need_rebuild?
+          return false if @context['type'] == 'dir'
           return false if @context['type'] == 'root' and @context['build_type'] == :dir_builder and not GitUtil.repo?(@context['full_build_path'])
           return false if @context['type'] == 'root' and @context['build_type'] == :git_direct_builder and GitUtil.repo?(@context['full_build_path'])
+          if @context['type'] == 'repo'
+            if @context['build_type'] == :dir_direct_builder
+              return false if GitUtil.repo?(@context['full_build_path'])
+            end
+          end
           log("Remove #{@context['full_build_path']}")
           FileUtils.rm_rf @context['full_build_path']
         end
